fix: bump gatsby and other deps
