+++ .gitignore
@@ -66,3 +66,4 @@ netlify.toml
 # mac FS
 .DS_Store
 .vscode/
+about.json
