--- manifest-config.js
@@ -1,11 +1,0 @@
-const colors = require('./colors');
-
-module.exports = {
-  name: 'Mate Gatsby Starter',
-  short_name: 'Mate Gatsby Starter',
-  start_url: '/',
-  background_color: colors.background,
-  theme_color: colors.primary,
-  display: 'minimal-ui',
-  icon: 'media/icon.png',
-};
