+++ package.json
@@ -4,8 +4,9 @@
   "description": "Gatsby v2 starter to create a top notch portfolio!",
   "main": "index.js",
   "scripts": {
-    "build": "gatsby clean && gatsby build",
-    "develop": "gatsby clean && gatsby develop",
+    "prepare": "node scripts/get-about",
+    "build": "yarn prepare && gatsby build",
+    "develop": "yarn prepare && gatsby develop",
     "start": "yarn develop",
     "serve": "yarn build && gatsby serve -p 9000",
     "lint": "eslint ./src",
@@ -36,15 +37,15 @@
     "@rebass/preset": "^4.0.5",
     "@tippy.js/react": "^3.1.1",
     "contentful": "^7.14.3",
-    "gatsby": "2.15.22",
+    "gatsby": "^2.23.3",
     "gatsby-plugin-google-analytics": "^2.3.2",
     "gatsby-plugin-manifest": "^2.4.3",
-    "gatsby-plugin-offline": "^3.2.7",
+    "gatsby-plugin-offline": "^3.2.9",
     "gatsby-plugin-react-helmet": "^3.3.1",
     "gatsby-plugin-styled-components": "^3.3.2",
     "gatsby-plugin-web-font-loader": "^1.0.4",
-    "gatsby-source-contentful": "2.3.12",
-    "gatsby-source-medium": "^2.3.2",
+    "gatsby-source-contentful": "^2.3.15",
+    "gatsby-source-medium": "^2.3.4",
     "gatsby-transformer-remark": "^2.8.13",
     "prop-types": "^15.6.2",
     "react": "^16.13.1",
@@ -63,11 +64,11 @@
     "@percy/cypress": "^2.1.1",
     "axe-core": "^3.5.4",
     "chalk": "^4.0.0",
-    "contentful-import": "^7.7.18",
+    "contentful-import": "^7.8.2",
     "cypress": "^4.7.0",
     "cypress-axe": "^0.8.1",
     "dotenv": "^8.2.0",
-    "envfile": "^5.2.0",
+    "envfile": "^6.2.0",
     "eslint": "^5.6.0",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-config-prettier": "^6.11.0",
